refactor: move chat http post to ChatService.
