+++ src/app/chatbot/chatbot.service.ts
@@ -27,33 +27,8 @@ export class ChatMessage {
 
 @Injectable({ providedIn: 'root' })
 export class ChatbotService {
-  private chatGptEndpoint = '/api/chat-gpt';
   private chatsEndpoint = '/api/chatbot/chats';
   private http = inject(HttpClient);
-
-  /**
-   * Sends a message to the chat-gpt endpoint.
-   * @param conversationHistory The conversation history.
-   * @returns A promise that resolves to the response from the chat-gpt endpoint.
-   */
-  async sendMessage(conversationHistory: ChatMessage[]): Promise<string> {
-    const payload = {
-      messages: conversationHistory.map((msg) => ({
-        role: msg.role,
-        content: msg.content
-      })),
-      model: 'gpt-3.5-turbo'
-    };
-    try {
-      const response = await firstValueFrom(
-        this.http.post<any>(`${this.chatGptEndpoint}`, payload)
-      );
-      return response.choices[0].message.content;
-    } catch (error) {
-      console.error('Error calling chat-gpt endpoint:', error);
-      throw error;
-    }
-  }
 
   getChats(runId: number, workgroupId: number): Observable<Chat[]> {
     return this.http.get<Chat[]>(`${this.chatsEndpoint}/${runId}/${workgroupId}`).pipe(
